move zk mock to mock dir
