+++ components/sequencer/zookeeper/zookeeper_sequencer_test.go
@@ -17,7 +17,7 @@ import (
 	"github.com/go-zookeeper/zk"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
-	"mosn.io/layotto/components/pkg/utils"
+	"mosn.io/layotto/components/pkg/mock"
 	"mosn.io/layotto/components/sequencer"
 	"mosn.io/pkg/log"
 	"testing"
@@ -38,7 +38,7 @@ func TestZookeeperSequencer_GetNextId(t *testing.T) {
 
 	//mock
 	ctrl := gomock.NewController(t)
-	client := utils.NewMockZKConnection(ctrl)
+	client := mock.NewMockZKConnection(ctrl)
 
 	path := "/" + key
 	var val int32 = 1
